Refactor: utils 폴더 생성 및 파일 이동
