+++ src/hooks/useModal.tsx
@@ -1,5 +1,15 @@
+import { useCallback } from 'react'
+
 const useModal = () => {
-  return <div>useModal</div>
+  const openModal = useCallback(() => {
+    //state setting
+  }, [])
+
+  const closeModal = useCallback(() => {
+    //state setting
+  }, [])
+
+  return { openModal, closeModal }
 }
 
 export default useModal
